Removed required claimant ID from all calls.
This is now part of the client: one client, one claimant ID by default.
It can be overridden during client creation, and each call that cares
about it also has an option to specify it. Only use case at present is
to allow qsvc to work using the plain EntroQ client (where it must set
the claimant ID from the request).