Remove final race from subq.
The race is a little subtle. It affects performance, not correctness, so
this is not a critical update.

The waiter, just after releasing the subq global mutex because it is
updating the channel map, can be preempted before entering the select.
If that happens, it will wait even though a notification came in after
it indicated it wanted to reserve the channel.

In other words, it declared its intent to listen, but because the select
hadn't started, the Notify default executes and the event is dropped on
the floor.

This fixes that case: the Notify function now loops on the select, and
if the default case is entered *and* there are reserved listeners, it
tries again until either the send succeeds or there are no more reserved
listeners.