subq Wait accepts a cond function and poll timeout.
The previous implementation of Wait admitted a macro-level race, where
the thing we want to poll (e.g., the presence of a claimable task) might
no longer be there after a notification arrives, but before we check the
channel again, a notification for it might get dropped.

This eliminates that race condition, in theory, because the call to cond
(which can be a call to TryClaim) happens inside the "reservation
window", where it is still indicated that something is listening on the
channel.

If cond takes a while, which it can if it's querying over the network,
notify has to wait for it, so that busy loop now includes sleeping.

The complexity went way up here. This may not be worth the trouble.

Also moved the large mapreduce test into the mem package. The small one
is still in there, and a small check test is, as well.