Add dependency error handler to worker, with tests.
Allows dependency errors to be upgraded to fatal in the worker, with
custom user logic.

Also allows dependency errors to be responded to (even if not upgraded).
For example, one might depend on a config task, and note that when there
is a dependency error on that task, it's time to reload it.