Remove in-memory implementation, switch back to REST.
The in-memory implementation might make a comeback at some point, but
there are many benefits to using the REST service, even though it's a
little slower.

1) It's easier to be flexible.

The number of ways to configure OPA is high, and we don't want to have
to explicitly support all of them. It's enough that EntroQ has to know
about authorization headers, now. There's no reason to add/force certain
kinds of permissions structures, monitoring, file reloading, grabbing
things from network services, etc.

2) OPA package splits (recent) make all of this simpler anyway.

The recent split of OPA config into packages that make overriding
functionality simpler also highlights the fact that it's better to not
have a mixed configuration.

Instead, we can provide example configuration files and some core
rules that understand the shape of the input data and the requirements
of the service itself (the kinds of results it needs), and we can
document how to change it if folks want to. The critical functionality
is located in basically two files (the queues utils and the authz
package), and every modification desired can be made in other packages
that are imported.

3) Kicking OPA to get it to pick up a new config is much less disruptive
   than kicking EntroQ.

Keeping those services and their lifecycles
separate is a compelling idea.