Fix long-standing move/retry-on-error bug with renewed tasks.
If the `RenewAllFor` function in `DoWithRenewAll` failed with an error,
it wiped out the `renewed` task list. This meant, for example, that
after a context cancellation, that list could be empty after a `Wait`.

If the context was canceled because, for example, the function to be
called during renewal output an error (such as MoveTaskError), and that
happened before the closure of the done channel, the renewed tasks would
be wiped out.

It was therefore a race in the select statement: either the select would
notice that it needed to exit gracefully first, or it would run the
renewal on a canceled context, which would fail. If the graceful exit
won the race, then everything was fine, but if the renewal function won
the race (to be immediately terminated because of cancelation), then the
renewed tasks were wiped out and error-based renewal would fail.