Fix (hopefully) subtle renewal race when using gRPC.
Also use `t.Run` to run subtests, allowing for calls like

`go test -run EQMemWorkerMoveOnError/case=wait-for-renewal`

This mitigates a possible renewal race when using gRPC, where the
context is canceled during an in-flight network call, and mutation
happens on the backend even though the client has lost connection with
it.

These races are not completely fixable *in general*, but this can happen
in common normal situations, because the context was being used for
basic goroutine management (rather than actual context cancelation from
the parent).

The new logic makes the renewal loop blocking, and separates out the
group context from the network call context, allowing network calls that
are in flight to finish unless canceled by the caller. Thus, an
in-flight renewal will delay return of the DoWithRenewAll function, even
though the function it is performing is already finished. This delay
should be small in normal operation, and is non-existent except in rare
cases.

Indeed, it required running the MoveOnError test many times to even get
the problem to appear. When it did appear, it was extremely confusing
and took a lot of effort to find.

The moral of the story is this: always know what your context is for.
Parents get to do whatever they want, but in your own functions, don't
mix "network call" contexts with "goroutine management" contexts unless
you are sure that the network call won't mess something up when canceled
just a tad too late to stop the call, but too early to get the response.