Simplify logic for move and retry errors.
All logic is now inside the worker function, none of it is embedded in
DoWithRenew* functions anymore. That approach was awfully hacky anyway,
and this approach is cleaner. The non-worker functions no longer need to
have any worker-based stuff in them at all. The move and retry errors
are captured inside the worker `Run` function and translated to nil
errors for the sake of the renewal function, allowing the renew return
to work properly.

Further, the worker error is no longer so muddy: we now keep track of
special cases in a separate variable, tested in a more sensible part of
the flow. This makes the whole thing easier to reason about and to keep
functionality separate.