Correctness fix for eqmem task ins w/ID but diff Q.
When a task with an ID specified is inserted, and SkipColliding is not
on, it was still possible to insert the task anyway if the queue
specified was different than the task that already existed in memory.
This can lead to duplicate tasks existing in different queues, which
violates an important invariant.

Added a test for this in InsertWithID, ensuring that even if a task is
inserted with ID into a completely different queue, a dependency error
is expected.

Verified that the test fails on the old code and passes on the new code.
Fixed the failing test that led to this discovery (pg was failing on
move, mem was not, turned out pg was right, the test was broken, but
fixing the test didn't leave me satisfied because of the divergence in
behavior).