Fixed race condition with queue locks. Showed up as tasks lost between insert and claim in a flaky SimpleWorker test.
Fixed the test output, as well, to make it clear what is happening.
Before this change, it wasn't clear that a task was getting dropped. Now
the output and the check functions make sense.

The subtlety in the main code was this:
- grab global lock, generate list of queue locks
- during that process, optionally generate new locks for formerly empty
  queues
- unlock
- grab queue locks, do work
- unlock
- grab global lock, clean up any empty queue locks

This last step broke things in some cases. A claim could come in and
grab the global lock, but cleanup already happened and it didn't know
about it. Furthermore, Modify was not reentrant because of this: an
insert operation could have its locks clobbered by a delete operation on
another routine (which is exactly what is happening with a claim/delete
worker pattern).

So Modify was called to insert something, and between creating the new
lock for an empty queue and doing its actual work, the worker would come
in with a Modify delete operation, clean up the lock on the empty queue,
and leave everything in a broken state.

Now we track who has *asked for* a lock, and drop that only after
unlocking it. This spans across mutex lock drops, and a goroutine
created with the lock watches for it to be ready to be cleaned up,
infrequently taking the global lock very briefly to check and perform
safe cleanup only on locks that nobody cares about anymore, not relying
on mutexes as the only signal for that status..